FlexMessageのハッシュを返すクラスとgiphyAPIを使うクラスの切り出し
